Restructure braces var sub to be powered by clear_scopes
This significantly simplifies the syntax, allows it to handle cases where the
string type is different at the end brace than the start brace, and only comes
with a minor limitation -- that successive closing braces after a variable
substitution will continue to clear_scopes and progressively wipe away all the
remaining meta_scope's that are active on that argument. This all goes away the
moment unquoted whitespace is encountered, and is not considered a major issue.