Add the two snippets which are obviously missing
"trap" and "wait"

Originally I was thinking that these didn't need snippets, and that's probably
still true. "trap" especially given that there isn't a meaningful default to put
in the block so use of this snippet barely saves any keystrokes. But, when all
the other builtins that use blocks are accounted for and these two aren't, it
seems like a mistake *not* to have them as well.