Drop the arbitrary distinction between ws_sep and param_sep
Having both of these was really a holdover from my sublime-fish package, where
it was useful because fish actually has other operators besides whitespace that
have the power to separate parameters. execline does not, but at first I wasn't
sure that would be the case because of blocks. Now it's clear that the
distinction is unnecessary because whitespace really is the only operator with
that power.